chore: category DTO 적용
