+++ src/controllers/index.ts
@@ -1,3 +1,6 @@
 export * from './user';
 export * from './candies';
 export * from './search';
+export * from './category';
+export * from './candy';
+
