+++ src/services/index.ts
@@ -1,3 +1,5 @@
 export * from './UserService';
 export * from './CandiesService';
 export * from './SearchService';
+export * from './CategoryService';
+export * from './CandyService';
